+++ Shmengine/sauce/renderer/vulkan_renderer/VulkanBackend.hpp
@@ -13,8 +13,8 @@ namespace Renderer::Vulkan
 
 	void on_resized(uint32 width, uint32 height);
 
-	bool32 begin_frame(float32 delta_time);
-	bool32 end_frame(float32 delta_time);
+	bool32 begin_frame(float64 delta_time);
+	bool32 end_frame(float64 delta_time);
 
 	void vk_render_target_create(uint32 attachment_count, Texture* const * attachments, Renderpass* pass, uint32 width, uint32 height, RenderTarget* out_target);
 	void vk_render_target_destroy(RenderTarget* target, bool32 free_internal_memory);
